+++ fastai/vision/models/unet.py
@@ -34,14 +34,14 @@ class UnetBlock(nn.Module):
         cat_x = self.relu(torch.cat([up_out, self.bn(s)], dim=1))
         return self.conv2(self.conv1(cat_x))
 
-class DynamicUnet(nn.Sequential):
+
+class DynamicUnet(SequentialEx):
     "Create a U-Net from a given architecture."
     def __init__(self, encoder:nn.Module, n_classes:int, blur:bool=False, blur_final=True,
                  self_attention:bool=False, sigmoid:bool=False, **kwargs):
         imsize = (256,256)
         sfs_szs = model_sizes(encoder, size=imsize)
         sfs_idxs = list(reversed(_get_sfs_idxs(sfs_szs)))
-        # TODO: add x-connection from input, then add extra output conv or resblock
         self.sfs = hook_outputs([encoder[i] for i in sfs_idxs])
         x = dummy_eval(encoder, imsize).detach()
 
@@ -63,6 +63,9 @@ class DynamicUnet(nn.Sequential):
 
         ni = x.shape[1]
         if imsize != sfs_szs[0][-2:]: layers.append(PixelShuffle_ICNR(ni, **kwargs))
+        layers.append(MergeLayer(dense=True))
+        ni += 3
+        layers.append(res_block(ni, **kwargs))
         layers += [conv_layer(ni, n_classes, ks=1, use_activ=False, **kwargs)]
         if sigmoid: layers.append(nn.Sigmoid())
         super().__init__(*layers)